changed events schema and moved event card to seperate component
