+++ cms/api/events/models/events.settings.json
@@ -2,7 +2,8 @@
 	"kind": "collectionType",
 	"collectionName": "events",
 	"info": {
-		"name": "events"
+		"name": "events",
+		"description": ""
 	},
 	"options": {
 		"increments": true,
@@ -29,8 +30,8 @@
 		"date": {
 			"type": "date"
 		},
-		"relevant_links": {
-			"type": "json"
+		"links": {
+			"type": "string"
 		}
 	}
 }
