+++ package.json
@@ -3,7 +3,7 @@
 	"description": "Our new website!",
 	"author": "Developer Student Club - KIIT",
 	"license": "MIT",
-	"version": "2.1.1",
+	"version": "2.2.0",
 	"private": true,
 	"scripts": {
 		"dev": "next dev --port=3333",
